requirements.txt
asgiref==3.4.1
astroid==2.9.3
attrs==21.4.0
autopep8==1.6.0
black==21.12b0
click==8.0.3
colorama==0.4.4
dj-database-url==0.5.0
Django==4.0.1
django-heroku==0.3.1
djangorestframework==3.13.1
drf-spectacular==0.21.1
gunicorn==20.1.0
inflection==0.5.1
isort==5.10.1
jsonschema==4.4.0
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
psycopg2==2.9.3
pycodestyle==2.8.0
pylint==2.12.2
pylint-django==2.5.0
pylint-plugin-utils==0.7
pyrsistent==0.18.1
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
tzdata==2021.5
uritemplate==4.1.1
whitenoise==5.3.0
wrapt==1.13.3
